feat(system): first advance — project management works, first ui version
First advance, the primary UI works, these are the main features.

---
# Project flow

- The program creates a **pseudo-installation** if never executed in the pc.
	- In the user home directory creates a directory called `.imglblr` and, inside that directory creates a file called `app-settings.json` this file will save the opened projects paths.
	- If the directory and the file already exists, it won't do anything.
- After the program installation is checked will do the following.
	- Open the `app-settings.json` file. This JSON file **contains data of the projects previously opened** by the user, it's `id` and `path`.
		- If there are project, will open an **editor window for each project**
		- If not, will show the **welcome window**

---
# UI

The UI contains the following views

- Welcome window
- Create project window
- Main window (editor)

### Welcome view

Create the `WelcomeWindow` class that shows the following:

- A Title with a welcome to the user and a text that describes the actions that the user can do
- Two buttons with the actions that the user can do.
	- Create new project will open the window to create a new project, this view will be described later.
	- Open project will open a file editor to select a project, after selected will open an editor in that path

### Create project view

This  `CreateProject` class is a view that shows the following

- A text input for the project name
- A text input for the project parent folder path
	- At the end of this input, have a button to open the file explorer and get the project parent path in a more user-friendly way
- A label that shows the final path of the project

### Main window view (editor)

This `MainWindow`  class contains the following view

- A menu bar for all the settings that the user can do

For now, this class is a little empty, but the components are build in an independent way.

## Components

The UI is made of multiple components to make the UI more escalable and reusable. These are the main components

### Menu bar

This component is the menu bar for the editor window. Contains the following options.

- File
	- New project
	- Close project
- View
	- Dataset viewer
	- Polygon editor
	- Open class editor
- Dataset
	- Load dataset
	- Load data
- Polygon Editor
	- Placeholder

This actions are placeholders for now and does not contain any logic inside them.

### Image viewer

This component takes an image path and show the image. This component will be shown in another component, the polygon editor

---
# Miscellaneous

## Project Manager

Also, a `ProjectManager` class exists, this class handle this handle all the previous opened projects, and it's used to open previous opened projects without open one by one in the welcome window

## Installer

This class is used to make the *pseudo-installation*

## Image loader

This loads the images and the dataset for the project.